% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activeSignaling_CellComm.R
\name{activeSignaling}
\alias{activeSignaling}
\title{Create gene regulator subnetworks based on gene signature value.}
\usage{
activeSignaling(data, sampTab, markers, grn, summary, fc = 5)
}
\arguments{
\item{data}{dgCMatrix; gene expression in each cell, found in
CellRouter@ndata.}

\item{sampTab}{data frame; the metadata information of the cells, found in
CellRouter@sampTab.}

\item{markers}{data frame; the gene signatures as calculated by the
findSignatures function of CellRouter.}

\item{grn}{data frame; the gene regulatory network table (GRN_table) as
calculated by the buildGRN method of CellRouter.}

\item{summary}{data frame; the summary of the files obtained in the
findpaths.simple function.}

\item{fc}{numeric; keep only vertex which signature value is smaller than the
fold changes.}
}
\value{
list; the pathways and the gene regulatory networks (gnrs). The
pathways contain the lists of genes that are the markers and the
regulators of the target populations for each interaction. The gnrs
contain the regultory sub network for each node in the summary and for each
target in the node.
}
\description{
Create subnetworks based on the gene signature value for each node in the
summary and for each target in the node.
}
