% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInfo.R
\name{addInfo}
\alias{addInfo}
\title{Add metadata information to CellRouter object.}
\usage{
addInfo(
  object,
  assay.type = "RNA",
  sample.name = "Sample1",
  metadata,
  colname,
  metadata.column = "population"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{metadata}{data frame or vector; metadata to be added.}

\item{colname}{character; column name to be added to \code{assay@sampTab}
in case metadata is a vector.}

\item{metadata.column}{character; column to selected from the metadata to be
added and included in \code{assay@sampTab}.}
}
\value{
CellRouter object with the sampTab slot updated.
}
\description{
Include metadata information in the CellRouter metadata table
\code{assay@sampTab}. If you have clusters identified by other algorithm,
you can use this function to perform all the analysis available in
CellRouter using your previously identified clusters: just select the
corresponding columns in the table \code{assay@sampTab}.
}
