% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparse_matrix.R
\name{apply1_sp}
\alias{apply1_sp}
\title{Apply for sparse matrix.}
\usage{
apply1_sp(X, FUN)
}
\arguments{
\item{X}{matrix; data.}

\item{FUN}{function.}
}
\value{
res; the result depends on the function. If the function aggregates,
then it will do it by rows and return a vector. If the function does not
aggreate, it will return a list with each row.
}
\description{
Sparse matrix version of the apply function. Avoid errors when working with
large matrix. Apply the function on each element or row.
}
