% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGRN_method.R
\docType{methods}
\name{buildGRN2}
\alias{buildGRN2}
\alias{buildGRN2,CellRouter-method}
\title{Predict a gene regulatory network.}
\usage{
buildGRN2(
  object,
  assay.type = "RNA",
  species = c("Hs", "Mm"),
  tfs = NULL,
  genes.use = NULL,
  zscore = 5,
  blocksize = 500,
  filename = "GRN.R"
)

\S4method{buildGRN2}{CellRouter}(
  object,
  assay.type = "RNA",
  species = c("Hs", "Mm"),
  tfs = NULL,
  genes.use = NULL,
  zscore = 5,
  blocksize = 500,
  filename = "GRN.R"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{species}{character; species: Hs for Homo Sapiens or Mm for Mus Musculus.}

\item{tfs}{character vector; transcription factors to use for GRN reconstruction.}

\item{genes.use}{character vector; genes to include in the gene regulatory
network. The default is to use all genes.}

\item{zscore}{numeric; zscore threshold to identify putative regulatory
interactions.}

\item{blocksize}{numeric; size of the blocks in which genes will be scaled.}

\item{filename}{character; filename where the GRN data will be saved.}
}
\value{
list; the GNR with the gene regulatory network graph, the GRN_table
with the gene regulatory network table, and the the tfs with the transcription
factors.
}
\description{
Predict a gene regulatory network allowing a customized list of transcription
factors or genes (tfs).
}
