% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKNN.R
\docType{methods}
\name{buildKNN}
\alias{buildKNN}
\alias{buildKNN,CellRouter-method}
\title{Build kNN graph.}
\usage{
buildKNN(
  object,
  assay.type = "RNA",
  k = 5,
  column.ann,
  num.pcs = 20,
  sim.type = "jaccard",
  filename = "graph_clusters.gml"
)

\S4method{buildKNN}{CellRouter}(
  object,
  assay.type = "RNA",
  k = 5,
  column.ann,
  num.pcs = 20,
  sim.type = "jaccard",
  filename = "graph_clusters.gml"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{k}{numeric; number of nearest neighbors to build a k-nearest neighbors
graph for trajectory reconstruction.}

\item{column.ann}{character; column in the metadata table specifying the
transitions to be identified. For example, if 'population' is provided,
transitions will be identified between clusters previously identified.
However, sorted cell populations or customized states can also be used.
Check our tutorials for detailed examples.}

\item{num.pcs}{numeric; number of principal components that will define the
space from where the kNN graph is identified. For example, if num.pcs = 10,
the kNN graph will be created from a 10-dimensional PCA space.}

\item{sim.type}{character; updates the kNN graph to encode cell-cell
similarities using the jaccard or invlog methods.}

\item{filename}{character; name of gml file containing the kNN graph.}
}
\value{
CellRouter object with the graph, ndata, and sampTab slots updated.
}
\description{
Create kNN graph using the PCA components and cell-cell similarities using
the jaccard or invlog methods. Write graph to file.
}
