% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\docType{methods}
\name{calculateCentroids}
\alias{calculateCentroids}
\alias{calculateCentroids,CellRouter-method}
\title{Calculate cluster centroids}
\usage{
calculateCentroids(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column,
  cluster.type = c("Cluster", "Subcluster")
)

\S4method{calculateCentroids}{CellRouter}(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column,
  cluster.type = c("Cluster", "Subcluster")
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.column}{character; the name of the column where the clustering
information is stored.}

\item{cluster.type}{character; the column where the clusters are
indicated. It could be either 'Cluster' or 'Subcluster'.}
}
\value{
list; the centroids dataframe for each tissue
sample. Each dataframe has the columns cluster, row, col, imagerow, imagecol.
The row and col are the centroid coordinates while the imagerow and imagecol
are the coordinates for plotting, as scaled in the bcs list.
}
\description{
Calculate spatial centroids for the clusters or the subclusters.
}
