% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\docType{methods}
\name{calculateDistanceMatrix}
\alias{calculateDistanceMatrix}
\alias{calculateDistanceMatrix,CellRouter-method}
\title{Calculate distance matrix}
\usage{
calculateDistanceMatrix(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.type = c("Cluster", "Subcluster"),
  spot_distance = 100,
  normalize = TRUE
)

\S4method{calculateDistanceMatrix}{CellRouter}(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.type = c("Cluster", "Subcluster"),
  spot_distance = 100,
  normalize = TRUE
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.type}{character; the column where the clusters are
indicated. It could be either 'Cluster' or 'Subcluster'.}

\item{spot_distance}{numeric; the distance between the center of the spots.
The default is 100 micrometers. If this parameters is 1, the distance
matrix will be relative to the rows and columns of the spots.}

\item{normalize}{boolean; normalize the distances dividing by the
total size of the tissue.}
}
\value{
the CellRouter object.
}
\description{
Calculate distance matrix for the clusters in each tissue sample.
}
