% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellComm_functions.R
\name{calculateLigandReceptor}
\alias{calculateLigandReceptor}
\title{Calculate Ligand and Receptors}
\usage{
calculateLigandReceptor(graph)
}
\arguments{
\item{graph}{igraph; the cell graph.}
}
\value{
dataframe; the dataframe with the cell pairs.
}
\description{
Calculate the ligand and receptor pairs from a graph. Used in the predict
interactions function.
}
