% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateObservedMean_CellComm.R
\name{calculateObservedMean}
\alias{calculateObservedMean}
\title{Mean expression between ligands and receptors.}
\usage{
calculateObservedMean(interactions, mean.expr)
}
\arguments{
\item{interactions}{data frame; the interactions as calculated by the
population.pairing function.}

\item{mean.expr}{list; the statistics of the expressed genes as calculated by
the computeValue method of CellRouter.}
}
\value{
data frame; the same as returned by the population.pairing function,
interactions with celltype1, celltype2, pair, ligand, and receptor, plus the
columns of mean ligand expression (mean.ligand), mean receptor expression
(mean.receptor), and mean pair expression (mean).
}
\description{
Calculate mean gene expression between ligand and receptor of each
interaction in the
}
