% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePvalue_CellComm.R
\name{calculatePvalue}
\alias{calculatePvalue}
\title{Calculate the p-value for the interactions.}
\usage{
calculatePvalue(p, nPerm = 100, interactions2)
}
\arguments{
\item{p}{list; the interaction data frame for the permutations calculated by
the clusterPermutation function.}

\item{nPerm}{numeric; the number of permutations.}

\item{interactions2}{data frame; the interactions as calculated by the
calculateObservedMean function, with the mean expression for ligands,
receptors and pairs.}
}
\value{
matrix?; the interactions for each pair, plus the pvalue column.
}
\description{
Calculate the rate of values which mean expression value for each
ligand-receptor pair is greater than the mean expression value for each
pairwise cell type.
}
