% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellComm_class.R
\docType{class}
\name{CellComm-class}
\alias{CellComm-class}
\alias{CellComm}
\alias{initialize,CellComm-method}
\alias{CellComm-initialize,initializeCellComm}
\title{The CellComm class.}
\usage{
\S4method{initialize}{CellComm}(.Object)
}
\arguments{
\item{.Object}{object.}
}
\value{
CellComm object.
}
\description{
An S4 class to represent the CellRouter object and its contents.

Initialize CellComm object with no stored data.
}
\details{
The objetct is used to analyse and find pathways of protein-protein
interaction connecting proteins in the surface to the proteins involved in
the transcriptional regulation.
}
\section{Slots}{

\describe{
\item{\code{expdata}}{data frame;}

\item{\code{ndata}}{data frame;}

\item{\code{sampTab}}{data frame;}

\item{\code{rdimension}}{data frame;}

\item{\code{graph}}{list;}

\item{\code{signatures}}{list;}

\item{\code{foldchange}}{list;}

\item{\code{sources}}{character;}

\item{\code{targets}}{character;}

\item{\code{directory}}{list;}

\item{\code{paths}}{data frame;}

\item{\code{networks}}{list;}

\item{\code{genes.trajectory}}{character vector;}

\item{\code{pathsinfo}}{list;}

\item{\code{dynamics}}{list;}

\item{\code{clusters}}{list;}

\item{\code{correlation}}{list;}

\item{\code{top.correlations}}{list;}

\item{\code{davidenrichment}}{list;}

\item{\code{pathwayenrichment}}{list;}
}}

