% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellnetworksPPI_CellComm.R
\name{cellnetworksPPI}
\alias{cellnetworksPPI}
\title{Create a weighted protein-protein interaction network based on cell-type specific correlation coefficients.}
\usage{
cellnetworksPPI(
  ppi,
  expr,
  samples,
  corThr = 0,
  column,
  dir.prefix = "results/paths"
)
}
\arguments{
\item{ppi}{igraph graph; the network as calculated by the createPPI function.}

\item{expr}{dgCMatrix; gene expression in each cell, found in
CellRouter@ndata.}

\item{samples}{data frame; the metadata information of the cells, found in
CellRouter@sampTab.}

\item{corThr}{numeric; correlation threshold.}

\item{column}{character; character; column in the metadata table to specify
how the data should be aggregated (e.g. per cell type, cluster or other
metadata available).}

\item{dir.prefix}{character; prefix of the directory where the files will be
saved.}
}
\value{
list; the networks (igraph objects) for each cell type/cluster/other
metadata.
}
\description{
Create a graph based on gene expression correlation for the protein-protein
interactions for each population. The edges are the intersection of the ones
in the protein-protein interaction network and the ones obtained from the
correlation of gene expression in the cells, the weights are the correlation
value. The edge lists are saved to files.
}
