% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRouter_class.R
\docType{class}
\name{CellRouter-class}
\alias{CellRouter-class}
\alias{CellRouter}
\title{The CellRouter class.}
\description{
An S4 class to represent the CellRouter object and its contents.
}
\details{
A multifaceted single-cell analysis platform that identifies complex
cell-state transition trajectories by using flow networks to explore the
subpopulation structure of multi-dimensional, single-cell omics data.
}
\section{Slots}{

\describe{
\item{\code{assays}}{list; different data used in the analysis.}

\item{\code{sampTab}}{data frame; store the information of the integrated analysis of
multiple assays.}

\item{\code{rdimension}}{data frame; cell embeddings from PCA.}

\item{\code{pca}}{list; data from principal component analysis.}

\item{\code{tsne}}{list; data from t-SNE dimensionality reduction.}

\item{\code{umap}}{list; data from UMAP dimensionality reduction.}

\item{\code{dr.custom}}{list; data from custom dimensionality reduction.}

\item{\code{dc}}{list; data from dc dimensionality reduction.}

\item{\code{custom}}{list; cell embeddings for custom space.}

\item{\code{var.genes}}{A character.}

\item{\code{graph}}{A list.}

\item{\code{signatures}}{A list.}

\item{\code{sources}}{A character.}

\item{\code{targets}}{A character.}

\item{\code{directory}}{list; directories of files obtained in the findPaths method.}

\item{\code{paths}}{data frame; paths between populations.}

\item{\code{networks}}{list; graphs between populations.}

\item{\code{genes.trajectory}}{character vector; genes considered in the trajectories.}

\item{\code{pathsinfo}}{list; information about the paths across populations.}

\item{\code{dynamics}}{list; the smoothed correlations between gene expression and
pseudotime in paths between populations.}

\item{\code{clusters}}{list; the clusters for the selected trajectories.}

\item{\code{correlation}}{list; correlations between gene expression and pseudotime
in paths between populations.}

\item{\code{top.correlations}}{list; gene trends (up or down) for each path.}

\item{\code{pathwayenrichment}}{list; pathway enrichment analysis on genes up or
down-regulated along trajectories.}

\item{\code{curves}}{list; the curves for the specified paths.}
}}

