% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{center_with_threshold}
\alias{center_with_threshold}
\title{Center the data using threshold.}
\usage{
center_with_threshold(center_data, threshold)
}
\arguments{
\item{center_data}{data frame; data to be processed.}

\item{threshold}{numeric; max value present in center_data.}
}
\description{
Helper function of plotSignaturesHeatmap, plotPathHeatmap, plotClusters...
}
