% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenesPseudotime_method.R
\name{clustergenes}
\alias{clustergenes}
\title{Helper function of clusterGenesPseudotime.}
\usage{
clustergenes(fits, num.clusters)
}
\arguments{
\item{fits}{matrix; gene expression by pseudotime.}

\item{num.clusters}{numerical; number of clusters.}
}
\description{
Cluster genes in num.clusters kinetic trends.
}
