% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGenesPseudotime_method.R
\docType{methods}
\name{clusterGenesPseudotime}
\alias{clusterGenesPseudotime}
\alias{clusterGenesPseudotime,CellRouter-method}
\title{Cluster kinetic profiles along trajectories.}
\usage{
clusterGenesPseudotime(object, num.clusters)

\S4method{clusterGenesPseudotime}{CellRouter}(object, num.clusters)
}
\arguments{
\item{object}{CellRouter object.}

\item{num.clusters}{numeric; number of clusters.}
}
\value{
CellRouter object with the clusters slot updated.
}
\description{
Cluster kinetic gene expression profiles along each trajectory into the
specified number of clusters using the Partitioning Around Medoids method.
}
