% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\name{cn_sampR_to_pattern}
\alias{cn_sampR_to_pattern}
\title{Helper function of ranked_findSpecGenes.}
\usage{
cn_sampR_to_pattern(sampR)
}
\arguments{
\item{sampR}{character vector; vector of cell populations.}
}
\value{
dgCMatrix matrix; ans_sparse_matrix.
}
\description{
Return a pattern for use in cn_testPattern (template matching). Convert the
cell population to matrix of 0 and 1, where the rows are the unique cell
populations and the columns are the population of each cell. Thus, the cell
values indicates the population of a cell.
}
