% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\name{cn_testPattern}
\alias{cn_testPattern}
\title{Helper function of ranked_findSpecGenes.}
\usage{
cn_testPattern(pattern, expDat)
}
\arguments{
\item{pattern}{pattern}

\item{expDat}{expression matrix}
}
\value{
data frame; row.names=geneids, pval=pval, cval=cval, holm=holm.
}
\description{
Get information about the line that represent each gene expression in each
cell population.
}
