% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{commToNames}
\alias{commToNames}
\title{Assign names to subpopulations.}
\usage{
commToNames(commObj, prefix)
}
\arguments{
\item{commObj}{igraph::multilevel.community; subpopulations.}

\item{prefix}{character; prefix.}
}
\description{
Assign names to subpopulations. Found in graphClustering. Helper function of
findClusters.
}
