% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTwoGroups.R
\docType{methods}
\name{compareTwoGroups}
\alias{compareTwoGroups}
\alias{compareTwoGroups,CellRouter-method}
\title{Compute fold changes and find gene signatures.}
\usage{
compareTwoGroups(
  object,
  assay.type = "RNA",
  column = "population",
  group1,
  group2,
  fc.threshold = 0.25
)

\S4method{compareTwoGroups}{CellRouter}(
  object,
  assay.type = "RNA",
  column = "population",
  group1,
  group2,
  fc.threshold = 0.25
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{column}{character; column in the metadata table to group cells for
differential expression. For example, if 'population' is specified,
population-specific gene signatures will be identified.}

\item{group1}{character; name of group 1 used in comparison.}

\item{group2}{character; name of group 2 used in comparison.}

\item{fc.threshold}{numeric; fold change threshold.}
}
\value{
data frame; a data frame with the mean expression of the genes in
group 1 (p column), the mean expression of the genes in group 2 (np column),
the fold changes (fc column), the p-value for the wilcox test (pv column),
and the adjusted p-value (p.adj column).
}
\description{
Compute mean expression of each gene in two populations and compares them
using fold changes, p-value of wilcox test and adjusted p-value using the
bonferroni method.
}
