% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDC.R
\name{computeDC}
\alias{computeDC}
\title{Dimensionality reduction using diffusion components.}
\usage{
computeDC(
  object,
  assay.type = "RNA",
  genes.use = NULL,
  k = 20,
  sigma = "local",
  seed = 1
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes.use}{character vector; genes used in diffusion components
dimensionality reduction. The default is all genes in normalized data.}

\item{k}{numeric; parameter k to be used by the DiffusionMap function from
destiny pakcage.}

\item{sigma}{character; parameter sigma to be used by the DiffusionMap
function from destiny pakcage.}

\item{seed}{numeric; random seed.}
}
\value{
CellRouter object with the dc slot updated.
}
\description{
Performs dimensionality reduction using the DiffusionMap function from the
destiny package. The destiny package is presenting errors in the installation,
so it is not imported in the description.
}
