% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\docType{methods}
\name{computeFC}
\alias{computeFC}
\alias{computeFC,CellRouter-method}
\title{Compute fold changes and find gene signatures.}
\usage{
computeFC(
  object,
  assay.type = "RNA",
  column = "population",
  pos.only = TRUE,
  min.pct = 0.1,
  fc.threshold = 0.25
)

\S4method{computeFC}{CellRouter}(
  object,
  assay.type = "RNA",
  column = "population",
  pos.only = TRUE,
  min.pct = 0.1,
  fc.threshold = 0.25
)
}
\arguments{
\item{object}{CellRouter object}

\item{assay.type}{character; the type of data to use.}

\item{column}{character; column in the metadata table to group cells for
differential expression. For example, if 'population' is specified,
population-specific gene signatures will be identified.}

\item{pos.only}{logical; only uses genes upregulated.}

\item{min.pct}{numeric; minimum percentage.}

\item{fc.threshold}{numerical; fold change threshold.}
}
\value{
CellRouter object with the signatures slot updated. This slot
contains a list with a dataframe for each group containing information about
mean.np, mean.p, fc (fold changes), pv (p-value), p.adj (p-value adjusted.).
}
\description{
Compute fold changes and find gene signatures in each group of cells.
}
