% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePCA_method.R
\docType{methods}
\name{computePCA}
\alias{computePCA}
\alias{computePCA,CellRouter-method}
\title{Perform principal component analysis.}
\usage{
computePCA(object, assay.type = "RNA", num.pcs, genes.use = NULL, seed = 1)

\S4method{computePCA}{CellRouter}(object, assay.type = "RNA", num.pcs, genes.use = NULL, seed = 1)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{num.pcs}{numeric; number of principal components to compute.}

\item{genes.use}{character vector; genes used in principal component
analysis. Default is all genes in normalized data.}

\item{seed}{numeric; seed.}
}
\value{
CellRouter object with the pca slot updated.
}
\description{
Compute the correlation between cells using principal component analysis and
plot the standard deviation graph.
}
