% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTSNE_method.R
\docType{methods}
\name{computeTSNE}
\alias{computeTSNE}
\alias{computeTSNE,CellRouter-method}
\title{Perform t-SNE.}
\usage{
computeTSNE(object, num.pcs, perplexity = 30, max_iter = 2000, seed = 7)

\S4method{computeTSNE}{CellRouter}(object, num.pcs, perplexity = 30, max_iter = 2000, seed = 7)
}
\arguments{
\item{object}{CellRouter object.}

\item{num.pcs}{numeric; number of principal components used for
dimensionality reduction using t-SNE.}

\item{perplexity}{numeric; perplexity.}

\item{max_iter}{numeric; maximum number of iterations.}

\item{seed}{numeric; seed.}
}
\value{
CellRouter object with the tsne slot updated.
}
\description{
Dimensionality reduction using t-SNE.
}
