% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUMAP.R
\docType{methods}
\name{computeUMAP}
\alias{computeUMAP}
\alias{computeUMAP,CellRouter-method}
\title{Perform UMAP.}
\usage{
computeUMAP(
  object,
  num.pcs,
  spread = 1,
  min_dist = 0.3,
  n_neighbors = 30,
  metric = "cosine",
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  repulsion_strength = 1,
  negative_sample_rate = 5,
  seed = 1
)

\S4method{computeUMAP}{CellRouter}(
  object,
  num.pcs,
  spread = 1,
  min_dist = 0.3,
  n_neighbors = 30,
  metric = "cosine",
  set_op_mix_ratio = 1,
  local_connectivity = 1,
  repulsion_strength = 1,
  negative_sample_rate = 5,
  seed = 1
)
}
\arguments{
\item{object}{CellRouter object.}

\item{num.pcs}{numeric; number of principal components used for
dimensionality reduction using UMAP.}

\item{spread}{numeric; used during automatic estimation of a/b parameters.}

\item{min_dist}{numeric; determines how close points appear in the final
layout.}

\item{n_neighbors}{numeric; number of nearest neighbors.}

\item{metric}{character or function; determines how distances between data
points are computed.}

\item{set_op_mix_ratio}{numeric in range [0,1]; determines who the knn-graph
is used to create a fuzzy simplicial graph.}

\item{local_connectivity}{numeric; used during construction of fuzzy
simplicial set.}

\item{repulsion_strength}{numeric; weighting applied to negative samples in
low dimensional embedding optimization.}

\item{negative_sample_rate}{numeric; determines how many non-neighbor points
are used per point and per iteration during layout optimization.}

\item{seed}{numeric; seed.}
}
\value{
CellRouter object with the umap slot updated.
}
\description{
Dimensionality reduction using UMAP provided by the umap package.
}
