% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeValue.R
\docType{methods}
\name{computeValueSubclusters}
\alias{computeValueSubclusters}
\alias{computeValueSubclusters,CellRouter-method}
\title{Compute expression statistics of each gene in each population.}
\usage{
computeValueSubclusters(
  object,
  assay.type = "RNA",
  genelist,
  column = "population",
  subcluster.column = "Subpopulation",
  clusters,
  fun = "max"
)

\S4method{computeValueSubclusters}{CellRouter}(
  object,
  assay.type = "RNA",
  genelist,
  column = "population",
  subcluster.column = "Subpopulation",
  clusters,
  fun = "max"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genelist}{character vector; genes to use in the analysis.}

\item{column}{character; column in the metadata table to group cells for
differential expression. For example, if 'population' is specified,
population-specific gene signatures will be identified.}

\item{subcluster.column}{character; the name of the column where the
subclustering information will be stored.}

\item{clusters}{character; selected clusters.}

\item{fun}{character; statistical function to summary the gene expression.}
}
\value{
list; the statistics of the expressed genes in the population (p
column), not in the population (np), and the percentage (percent).
}
\description{
Compute maximum, median or mean expression of each gene in each population.
}
