% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayenrichment.R
\name{convertID}
\alias{convertID}
\title{Helper function of convertIDs.}
\usage{
convertID(ids, ens, from, to)
}
\arguments{
\item{ids}{list; table containing mappings between gene identifiers.}

\item{ens}{character; gene name.}

\item{from}{character; gene identifier.}

\item{to}{character; gene identifier.}
}
\value{
list; gene list.
}
\description{
Convert gene identifiers.
}
