% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayenrichment.R
\name{convertIDs}
\alias{convertIDs}
\title{Helper function of pathwayenrichment.}
\usage{
convertIDs(ids, geneList, from, to)
}
\arguments{
\item{ids}{list; table containing mappings between gene identifiers.}

\item{geneList}{character vector; gene list to be converted.}

\item{from}{character; gene identifier.}

\item{to}{character; gene identifier.}
}
\value{
list; gene list.
}
\description{
Convert gene lists from one identifier to another.
}
