% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPseudotime_method.R
\docType{methods}
\name{correlationPseudotime}
\alias{correlationPseudotime}
\alias{correlationPseudotime,CellRouter-method}
\title{Compute the correlation of each gene along the trajectories.}
\usage{
correlationPseudotime(object, type = c("slope", "spearman", "pearson"))

\S4method{correlationPseudotime}{CellRouter}(object, type = c("slope", "spearman", "pearson"))
}
\arguments{
\item{object}{CellRouter object.}

\item{type}{character; correlation method: linear regression (slope),
spearman or pearson.}
}
\value{
CellRouter object with the correlation slot updated.
}
\description{
Computes the correlation between the gene expression and the pseudotime of
for each trajectory using a specified method.
}
