% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cRamp_helper.R
\name{cRamp2}
\alias{cRamp2}
\title{Convert to values to 'white', "goldenrod1", and "darkorange2" colors.}
\usage{
cRamp2(x)
}
\arguments{
\item{x}{values to be converted to colors}
}
\description{
Helper function.
}
