% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cRamp_helper.R
\name{cRampClust}
\alias{cRampClust}
\title{Convert values to colors.}
\usage{
cRampClust(x, num_colors)
}
\arguments{
\item{x}{Values to be converted to columns.}

\item{num_colors}{Number of colors to interpolate.}
}
\description{
Helper function of addInfo and findClusters.
}
