% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRouter_class.R
\name{CreateCellRouter}
\alias{CreateCellRouter}
\title{Create CellRouter.}
\usage{
CreateCellRouter(
  rawdata,
  assay.type = "RNA",
  min.genes,
  min.cells,
  is.expr = 0
)
}
\arguments{
\item{rawdata}{AnyMatrix; raw data, cells should be represented as columns,
genes should be represented as rows.}

\item{assay.type}{character; the type of data provided to the CellRouter
object.}

\item{min.genes}{numeric; minimum of genes a cell should express for it to be
considered.}

\item{min.cells}{numeric; minimum of cells in which a gene should be
expressed for it to be considered.}

\item{is.expr}{numeric; the minimum expression value a gene should have to be
considered expressed.}
}
\description{
Create the object and check its parameters.
}
