% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathsAndNetworks.R
\name{createNetworksFromPaths}
\alias{createNetworksFromPaths}
\title{Create networks from paths.}
\usage{
createNetworksFromPaths(summary, paths, n.targets, data, sampTab, top.paths)
}
\arguments{
\item{summary}{summary}

\item{paths}{paths}

\item{n.targets}{number of targets}

\item{data}{data}

\item{sampTab}{metadata}

\item{top.paths}{top paths}
}
\value{
networks
}
\description{
Create and return networks.
}
