% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPathsAndNetworks.R
\name{createPaths}
\alias{createPaths}
\title{Create paths.}
\usage{
createPaths(summary, paths, data, sampTab, top.paths)
}
\arguments{
\item{summary}{summary}

\item{paths}{paths}

\item{data}{data}

\item{sampTab}{metadata table}

\item{top.paths}{top paths}
}
\value{
network
}
\description{
Create paths and return network.
}
