% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTFNetwork.R
\name{createTFNetwork}
\alias{createTFNetwork}
\title{Create transcription factor network.}
\usage{
createTFNetwork(
  object,
  assay.type = "RNA",
  apathways,
  grn.data,
  interactions,
  genelist,
  celltype,
  column = "celltype5"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{apathways}{list; the pathways as calculated by the activeSignaling
function.}

\item{grn.data}{list; the gene regulatory network as created by the buildGRN
function.}

\item{interactions}{character; the interactions considered in the analysis.}

\item{genelist}{character vector; the genes considered in the analysis.}

\item{celltype}{character; the cell type.}

\item{column}{character; column corresponding to the cell type.}
}
\value{
list; the igraph transcription factor network and the plot.
}
\description{
Create and plot the transcription factor network for the selected interaction
and genes. Save plot to file. Load the ggnetwork package before using this
function.
}
