% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCurves.R
\name{curves_info}
\alias{curves_info}
\title{Helper function of processCurves.}
\usage{
curves_info(
  object,
  assay.type = "RNA",
  num.cells,
  neighs,
  column.ann = "community",
  column.color = "colors"
)
}
\arguments{
\item{object}{CellRouter object}

\item{assay.type}{character; the type of data to use.}

\item{num.cells}{numeric; number of cells.}

\item{neighs}{numeric; neighs.}

\item{column.ann}{character; column corresponding to the population.}

\item{column.color}{character; column corresponding to the color.}
}
\value{
list; list with distr, path, pathInfo, pseudotime for each path, and
path_data.
}
\description{
Get information about the paths between populations. This is the version of
pathsinfo for the curves found in findCurves.
}
