% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSpace_method.R
\docType{methods}
\name{customSpace}
\alias{customSpace}
\alias{customSpace,CellRouter-method}
\title{Customized coordinates used to visualize data in a reduced dimension space.}
\usage{
customSpace(object, matrix)

\S4method{customSpace}{CellRouter}(object, matrix)
}
\arguments{
\item{object}{CellRouter object.}

\item{matrix}{matrix; cell embeddings.}
}
\value{
CellRouter object with the custom slot updated.
}
\description{
Attribute a matrix to \code{object@custom} as a list.
}
