% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_method.R
\docType{methods}
\name{dotplot}
\alias{dotplot}
\alias{dotplot,CellRouter-method}
\title{Dot plot of genes in populations.}
\usage{
dotplot(
  object,
  assay.type = "RNA",
  genes.use,
  column = "population",
  thr,
  order
)

\S4method{dotplot}{CellRouter}(
  object,
  assay.type = "RNA",
  genes.use,
  column = "population",
  thr,
  order
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes.use}{character vector; selected genes.}

\item{column}{character; column name in cellrouter@sampTab indicating the
population.}

\item{thr}{numeric; threshold.}

\item{order}{character vector; ordered populations to be considered.}
}
\value{
ggplot2; plot.
}
\description{
Present a dot plot of selected genes in the specified populations and save to
file. The dot size represents the percentage of cells expressing the
indicated genes in each cluster and the dot color intensity represents the
average expression level of the indicated genes.
}
