% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_method.R
\docType{methods}
\name{dotplotByMetadata}
\alias{dotplotByMetadata}
\alias{dotplotByMetadata,CellRouter-method}
\title{Dot plot of genes in populations.}
\usage{
dotplotByMetadata(
  object,
  assay.type = "RNA",
  genes.use,
  column = "population",
  column2,
  thr,
  order,
  order2,
  cols
)

\S4method{dotplotByMetadata}{CellRouter}(
  object,
  assay.type,
  genes.use,
  column,
  column2,
  thr,
  order,
  cols
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes.use}{character vector; genes to show.}

\item{column}{character; column name in cellrouter@sampTab indicating the
population.}

\item{column2}{character; column name in cellrouter@sampTab indicating the
groups that the population form column belongs.}

\item{thr}{numeric; threshold.}

\item{order}{character vector; ordered populations from column to be
considered.}

\item{order2}{character vector; ordered populations from column2 to be
considered.}

\item{cols}{numeric; number of columns in the output figure.}
}
\value{
ggplot2; plot.
}
\description{
Present a dot plot of selected genes in the specified populations in the
column argument by the ones in the column2 argument and save to file. The
dot size represents the percentage of cells expressing the indicated genes in
each cluster and the dot color intensity represents the average expression
level of the indicated genes.
}
