% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotSignatureScore.R
\docType{methods}
\name{dotplotSignatureScore}
\alias{dotplotSignatureScore}
\alias{dotplotSignatureScore,CellRouter-method}
\title{Dot plot signature score.}
\usage{
dotplotSignatureScore(object, assay.type = "RNA", genes.use)

\S4method{dotplotSignatureScore}{CellRouter}(object, assay.type = "RNA", genes.use)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes.use}{character vector; genes to show.}
}
\value{
ggplot2; plot.
}
\description{
Present a dot plot of selected genes in all populations in
sampTab@population and save to file.
}
