% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGRN_method.R
\name{extractRegulators}
\alias{extractRegulators}
\title{Extract transcriptional regulators.}
\usage{
extractRegulators(zscores, corrMatrix, genes, threshold)
}
\arguments{
\item{zscores}{matrix; zscores.}

\item{corrMatrix}{matrix; correlation matrix.}

\item{genes}{character vector; genes.}

\item{threshold}{numeric; threshold.}
}
\description{
Helper function globalGNR.
}
