% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{filterCells}
\alias{filterCells}
\title{Quality control.}
\usage{
filterCells(
  object,
  assay.type = "RNA",
  variables,
  thresholds.low,
  thresholds.high
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{variables}{character list; column names in assay@sampTab of the
variables the cells will be filtered on, such as number of detected genes or
mitochondrial content.}

\item{thresholds.low}{numeric vector; cells with values lower than the ones provided
here are filtered out.}

\item{thresholds.high}{numeric vector; cells with values higher than the ones
provided here are filtered out.}
}
\value{
CellRouter object with the rawdata, ndata, and sampTab slots updated.
}
\description{
Filter out cells based on variables provided in the parameter variables.
}
