% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCurves_method.R
\name{find_curve}
\alias{find_curve}
\title{Find curve}
\usage{
find_curve(
  object,
  metadata,
  column,
  population_sequence,
  dims.use,
  reduction.type
)
}
\arguments{
\item{object}{CellRouter object.}

\item{metadata}{data frame; metadata table with cell annotations.}

\item{column}{character; column in the metadata table specifying whether
transitions are between clusters or other annotations, such as sorted
populations.}

\item{population_sequence}{list of character vectors; the ordered sequence
of the populations that will be used to calculate the curve.}

\item{dims.use}{numeric vector; the number of dimensions to use.}

\item{reduction.type}{character; the dimension reduction space to be used:
pca, tsne, DC of diffusion components, umap, or custom.}
}
\value{
data frame; the cells, their distance from the origin of the curve,
and the curve they belong.
}
\description{
Find pseudo-time curve using the princurve library, given the gene path.
Alternative to the findPaths method.
}
