% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCurves_method.R
\name{find_sequence}
\alias{find_sequence}
\title{Find population sequence}
\usage{
find_sequence(object, assay.type = "RNA", column, source, target)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{column}{character; column in the metadata table specifying whether
transitions are between clusters or other annotations, such as sorted
populations.}

\item{source}{character; name of the source population.}

\item{target}{character; name of the target population.}
}
\value{
character vector; the populations, including the source and target
ones.
}
\description{
Find the populations between the source and target ones.
}
