% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGRN_method.R
\name{find_tfs}
\alias{find_tfs}
\title{Find transcription factors.}
\usage{
find_tfs(species = "Hs")
}
\arguments{
\item{species}{character; species: Hs for human or Mm for mouse.}
}
\description{
Identify transcriptional regulators based on gene ontology annotations.
Helper function of buildGRN.
}
