% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\docType{methods}
\name{findClusters}
\alias{findClusters}
\alias{findClusters,CellRouter-method}
\title{Identify clusters.}
\usage{
findClusters(
  object,
  assay.type = "RNA",
  sample.name = "Sample1",
  method = "graph.clustering",
  k = 20,
  num.pcs = 20,
  nn.type = "nng",
  sim.type = "jaccard"
)

\S4method{findClusters}{CellRouter}(
  object,
  assay.type = "RNA",
  sample.name = "Sample1",
  method = "graph.clustering",
  k = 20,
  num.pcs = 20,
  nn.type = "nng",
  sim.type = "jaccard"
)
}
\arguments{
\item{object}{CellRouter object}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{method}{character; graph.clustering or model.clustering.}

\item{k}{numeric; number of nearest neighbors to build a k-nearest neighbors
graph.}

\item{num.pcs}{numeric; number of principal components that will define the
space from where the kNN graph is identified. For example, if num.pcs = 10,
the kNN graph will be created from a 10-dimensional PCA space.}

\item{nn.type}{character; method to find the k-nearest neighbor graph. If
'nng', the code will use the nng function of the cccd package. If 'knn' or
'snn', it will use the functions from the bluster package, which are
recommended for big data due to their efficiency.}

\item{sim.type}{character; updates the kNN graph to encode cell-cell
similarities using the jaccard or invlog methods.}
}
\value{
CellRouter object with the slots updated according to the chosen
method: graph.clustering updates graph, rawdata, ndata, and sampTab;
model.clustering updates sampTab.
}
\description{
Identify clusters based on graph-clustering or model based clustering. For
model based clustering it is necessary to load the mclust package manually.
}
