% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCurves_method.R
\docType{methods}
\name{findCurves}
\alias{findCurves}
\alias{findCurves,CellRouter-method}
\title{Find pseudo-time curves.}
\usage{
findCurves(
  object,
  assay.type = "RNA",
  column = "population",
  population_sequences = list(),
  find_populations = FALSE,
  sources = NULL,
  targets = NULL,
  dims.use,
  reduction.type
)

\S4method{findCurves}{CellRouter}(
  object,
  assay.type = "RNA",
  column = "population",
  population_sequences = list(),
  find_populations = FALSE,
  sources = NULL,
  targets = NULL,
  dims.use,
  reduction.type
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{column}{character; column in the metadata table specifying whether
transitions are between clusters or other annotations, such as sorted
populations.}

\item{population_sequences}{list of character vectors; the ordered sequence
of the populations that will be used to calculate the curve.}

\item{find_populations}{boolean; if TRUE, the population sequence will be
found by the algorithm given a source and target population, if FALSE,
the population sequence will be specified by the user.}

\item{sources}{character vector; name of the source population.}

\item{targets}{character vector; name of the target population.}

\item{dims.use}{numeric vector; the number of dimensions to use.}

\item{reduction.type}{character; the dimension reduction space to be used:
pca, tsne, DC of diffusion components, umap, or custom.}
}
\value{
CellRouter object with the curves slot updated.
}
\description{
Find pseudo-time curves using the princurve library. Alternative to the
findPaths method. Add to CellRouter@curves a list with the cells, their
distance from the origin of the curve, and the population they belong,
for each path curve.
}
