% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\name{findmarkers}
\alias{findmarkers}
\title{Helper function of findSignatures.}
\usage{
findmarkers(object)
}
\arguments{
\item{object}{CellRouter object.}
}
\value{
data frame;
}
\description{
Find markers for each cell population. Create a table of genes,
fc_subpopulation, subpopulation with max expression.
}
