% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPaths_method.R
\docType{methods}
\name{findPaths}
\alias{findPaths}
\alias{findPaths,CellRouter-method}
\title{Identify trajectories connecting source and target populations.}
\usage{
findPaths(
  object,
  assay.type = "RNA",
  sources,
  targets,
  column = "population",
  maindir,
  method
)

\S4method{findPaths}{CellRouter}(
  object,
  assay.type = "RNA",
  sources,
  targets,
  column = "population",
  maindir,
  method
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sources}{character vector; names of the source populations.}

\item{targets}{character vector; names of the target populations.}

\item{column}{character; column in the metadata table specifying whether
transitions are between clusters or other annotations, such as sorted
populations.}

\item{maindir}{character; directory where the files will be saved.}

\item{method}{character; method used to determine the source and target cells
based on the source and target populations.}
}
\value{
CellRouter object with the directory slot updated.
}
\description{
Identify trajectories based on euclidean, maximum, manhattan, canberra,
binary or graph network. Writes to the computer terminal calling a java code
and that saves the data in files which address is stored in assay@sampTab.
These files will be processed by the processTrajectories function. The java
code is written in the CellRouterJava folder in the package directory. The
java folder in the package directory has the code with some modifications to
be used in the findPathsRJava function.
}
