% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpaths.simple_CellComm.R
\docType{methods}
\name{findpaths.simple}
\alias{findpaths.simple}
\alias{findpaths.simple,CellComm-method}
\title{Identify trajectories connecting populations in the interaction.}
\usage{
findpaths.simple(sources.targets, file, dir, maindir)

\S4method{findpaths.simple}{CellComm}(sources.targets, file, dir, maindir)
}
\arguments{
\item{sources.targets}{list; the sources and targets of the interaction
receptor as calculated by the sourcestargets function:
\code{sts[['receptor']][[interaction]]}.}

\item{file}{character; the .txt file corresponding to the cell population. ?}

\item{dir}{character; directory for the selected interaction.}

\item{maindir}{character; directory where the directories for each
interaction will be saved.}
}
\value{
Message indicating conclusion.
}
\description{
Identify signaling pathway source and target genes (cell surface receptors
and transcriptional regulators, specifically) in the weighted protein
interaction network. Create a .sh file and run it using the terminal to call
the java code in Flow.jar.
The computations are stored in the specified directories and can
be read using the summarize function.
The java code is written in the CellCommJava folder in the package directory.
The java folder in the package directory has the code with some
modifications to be used in the findpaths.simpleRJava function.
}
