% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpaths.simpleRJava_CellComm.R
\docType{methods}
\name{findpaths.simpleRJava}
\alias{findpaths.simpleRJava}
\alias{findpaths.simpleRJava,CellComm-method}
\title{Identify signaling pathways connecting cell surface receptors to transcriptional regulators in the protein-protein interaction network.}
\usage{
findpaths.simpleRJava(object, sources.targets, file, dir, maindir)

\S4method{findpaths.simpleRJava}{CellComm}(object, sources.targets, file, dir, maindir)
}
\arguments{
\item{object}{CellComm object.}

\item{sources.targets}{list; the sources and targets of the interaction
receptor as calculated by the sourcestargets function:
\code{sts[['receptor']][[interaction]]}.}

\item{file}{character; the .txt file corresponding to the cell population. ?}

\item{dir}{character; directory for the selected interaction.}

\item{maindir}{character; directory where the directories for each
interaction will be saved.}
}
\value{
Message indicating conclusion.
}
\description{
Identify trajectories surface receptors to transcriptional regulators in the
protein-protein interaction network. The code uses the rJava library and does
not write and run a .sh file as the findpaths.simple function does.
The java code saves the data in files which address is stored in
assay@sampTab. These files will be processed by the processTrajectories
function. This function uses the java code stored in the java folder in the
package directory. If you are a linux user and the rJava
library is not installed, please check the topic at
\url{https://stackoverflow.com/questions/12872699/error-unable-to-load-installed-packages-just-now}.
}
