% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathsRJava.R
\docType{methods}
\name{findPathsRJava}
\alias{findPathsRJava}
\alias{findPathsRJava,CellRouter-method}
\title{Identify trajectories connecting source and target populations.}
\usage{
findPathsRJava(
  object,
  assay.type = "RNA",
  sources,
  targets,
  column = "population",
  maindir,
  method
)

\S4method{findPathsRJava}{CellRouter}(
  object,
  assay.type = "RNA",
  sources,
  targets,
  column = "population",
  maindir,
  method
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sources}{character vector; names of the source populations.}

\item{targets}{character vector; names of the target populations.}

\item{column}{character; column in the metadata table specifying whether
transitions are between clusters or other annotations, such as sorted
populations.}

\item{maindir}{character; directory where the files will be saved.}

\item{method}{character; method used to determine the source and target cells
based on the source and target populations. The available methods are
'euclidean', 'maximum', 'manhattan', 'canberra', 'binary', 'graph'.}
}
\value{
CellRouter object with the directory slot updated.
}
\description{
Identify trajectories based on euclidean, maximum, manhattan, canberra,
binary or graph network. The code uses the rJava library and does not
communicate with the terminal as the findPaths function does. The
computations are stored in the specified directories and can be read using
the summarize function. This function uses the java code stored in the java
folder in the package directory. If you are a linux user and the rJava
library is not installed, please check the topic at
\url{https://stackoverflow.com/questions/12872699/error-unable-to-load-installed-packages-just-now}.
}
