% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\name{findSignatures}
\alias{findSignatures}
\title{Identify gene signatures.}
\usage{
findSignatures(
  object,
  assay.type = "RNA",
  column = "population",
  test.use = "wilcox",
  pos.only = TRUE,
  min.pct = 0.1,
  fc.threshold = 0.25,
  fc.tm = FALSE
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{column}{character; specify the groups to compare.}

\item{test.use}{character; differential expression test to use. Alternative
is based on template-matching. Possible values: wilcox or template.}

\item{pos.only}{logical; use only upregulated genes.}

\item{min.pct}{numeric; minimum percentage.}

\item{fc.threshold}{numerical; fold change threshold.}

\item{fc.tm}{logical; whether to include fold change values in the
template-matching differential expression analysis.}
}
\value{
data frame; the population, fold change (fc), mean p, mean np,
p-value, p-value adjusted, and gene.
}
\description{
Identify gene signatures using the wilcoxon test or template-matching
methods.
}
