% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\docType{methods}
\name{findSubclusters}
\alias{findSubclusters}
\alias{findSubclusters,CellRouter-method}
\title{Find subclusters}
\usage{
findSubclusters(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column = "population",
  subcluster.column = "Subpopulation",
  method = c("mclust", "dbscan"),
  eps = 2,
  minPts = 3
)

\S4method{findSubclusters}{CellRouter}(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column = "population",
  subcluster.column = "Subpopulation",
  method = c("mclust", "dbscan"),
  eps = 2,
  minPts = 3
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.column}{character; the name of the column where the
clustering information is stored.}

\item{subcluster.column}{character; the name of the column where the
subclustering information will be stored.}

\item{method}{character; method to perform clustering (mclust or dbscan).}

\item{eps}{numeric; eps parameter for the DBSCAN method.}

\item{minPts}{numeric; minPts parameter for the DBSCAN method.}
}
\value{
the CellRouter object.
}
\description{
Calculate spatial clusters inside the clusters already identified by
gene expression information.
}
