% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{FindVariableGenes}
\alias{FindVariableGenes}
\title{Find Variable Genes}
\usage{
FindVariableGenes(
  object,
  assay.type = "RNA",
  method = "vst",
  pvalue = 1e-06,
  loess.span = 0.3
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.]}

\item{method}{character; the method to perform the identification of variable
genes: coefficient_variation or vst (more efficient).}

\item{pvalue}{numeric; the p-value threshold for the genes to be considered.
Argument for the coefficient_variation method.}

\item{loess.span}{numeric; argument for the vst method.}
}
\value{
dataframe; the information about the genes.
}
\description{
Preprocessing step to find variable genes to use in the downstream analysis.
It reduces the number of genes to be considered and speeds up the workflow.
It also plots the genes by their p-value.
}
