% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellRouter_class.R
\docType{class}
\name{FuscaAssay-class}
\alias{FuscaAssay-class}
\alias{FuscaAssay}
\title{CellRouter assay.}
\description{
Assay to create CellRouter.
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{AnyMatrix; raw data, cells should be represented as columns,
genes should be represented as rows.}

\item{\code{ndata}}{AnyMatrix; original data, cells should be represented as columns,
genes should be represented as rows.}

\item{\code{scale.data}}{AnyMatrix; scaled data.}

\item{\code{sampTab}}{dataframe; data frame; metadata information with cells names as
row names, number of genes as nGene, number of UMI as nUMI, and cell names as
conditions. It also includes the clusters and subclusters of the data in the
assay.}

\item{\code{image}}{list; image information of the ST assay.}
}}

