% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGRN_method.R
\name{globalGRN}
\alias{globalGRN}
\title{Create global gene regulatory network.}
\usage{
globalGRN(expr, tfs, threshold, blocksize)
}
\arguments{
\item{expr}{dgCMatrix; gene expression in each cell, found in
CellRouter@ndata.}

\item{tfs}{character vector; transcription factors.}

\item{threshold}{numeric; zscore threshold to identify regulatory
interactions.}

\item{blocksize}{numeric; size of the blocks in which genes will be scaled.}
}
\description{
Helper function of buildGRN.
}
