% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\docType{methods}
\name{graphClustering}
\alias{graphClustering}
\alias{graphClustering,CellRouter-method}
\title{Graph-based clustering.}
\usage{
graphClustering(
  object,
  assay.type = "RNA",
  k = 5,
  num.pcs,
  nn.type = "nng",
  sim.type = "jaccard",
  filename = "graph_subpopulations.gml"
)

\S4method{graphClustering}{CellRouter}(
  object,
  assay.type = "RNA",
  k = 5,
  num.pcs,
  nn.type = "nng",
  sim.type = "jaccard",
  filename = "graph_subpopulations.gml"
)
}
\arguments{
\item{object}{CellRouter object}

\item{assay.type}{character; the type of data to use.}

\item{k}{numeric; number of nearest neighbors to build a k-nearest neighbors
graph.}

\item{num.pcs}{numeric; number of principal components that will define the
space from where the kNN graph is identified. For example, if num.pcs = 10,
the kNN graph will be created from a 10-dimensional PCA space.}

\item{nn.type}{character; method to find the k-nearest neighbor graph. If
'nng', the code will use the nng function of the cccd package. If 'knn' or
'snn', it will use the functions from the bluster package, which are
recommended for big data due to their efficiency.}

\item{sim.type}{character; updates the kNN graph to encode cell-cell
similarities using the jaccard or invlog methods.}

\item{filename}{character; save .gml file containing the kNN graph.}
}
\value{
CellRouter object with the graph, rawdata, ndata, and sampTab slots
updated.
}
\description{
Identify clusters using graph-based model. Use jaccard or invlog similarity.
Update graph, rawdata, ndata and sampTab.
}
