% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnScoresHeatmap.R
\docType{methods}
\name{grnScoresHeatmap}
\alias{grnScoresHeatmap}
\alias{grnScoresHeatmap,CellRouter-method}
\title{Compute and plot heatmap of GRN scores for transcriptional regulators.}
\usage{
grnScoresHeatmap(
  object,
  ggrn,
  tfs,
  transitions,
  direction = c("up", "down", "both"),
  dir.targets = c("up", "down"),
  q.up = 0.95,
  q.down = 0.05,
  flip
)

\S4method{grnScoresHeatmap}{CellRouter}(
  object,
  ggrn,
  tfs,
  transitions,
  direction = c("up", "down", "both"),
  dir.targets = c("up", "down"),
  q.up = 0.95,
  q.down = 0.05,
  flip
)
}
\arguments{
\item{object}{CellRouter object.}

\item{ggrn}{igraph; the gene regulatory network.}

\item{tfs}{character vector; gene names of transcriptional regulators.}

\item{transitions}{character vector; selected transitions of interest.}

\item{direction}{character; plot genes up-regulated, down-regulated or both
along trajectories.}

\item{dir.targets}{character; whether the predicted targets are up or
down-regulated.}

\item{q.up}{numeric; cutoff to select top q.up transcriptional regulators.}

\item{q.down}{numeric; cutoff to select top q.down transcriptional regulators.}

\item{flip}{boolean; apply coordinate flip (horizontal and vertical) to the
plot.}
}
\value{
ggplot2; plot.
}
\description{
Compute and plot heatmap of the GRN scores of most representative genes in
trajectories and save to file. Integrate gene regulatory networks with gene
expression dynamics along the trajectories to calculate a GRN score and
identify putative regulators of these cell-fate transitions.
}
