% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGRN_method.R
\name{ig_tabToIgraph}
\alias{ig_tabToIgraph}
\title{Create an igraph object for the gene regulatory network.}
\usage{
ig_tabToIgraph(grnTab, simplify = FALSE, directed = FALSE, weights = TRUE)
}
\arguments{
\item{grnTab}{matrix; table containing gene regulatory relationships. Table
of TF, TF, maybe zscores, maybe correlations.}

\item{simplify}{boolean; simplify the graph. For example, remove loops.}

\item{directed}{boolean; create a directed or undirected graph.}

\item{weights}{boolean; add gene correlations as weights of the graph.}
}
\value{
iGraph object.
}
\description{
Helper function of buildGRN.
}
