% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionmatrix_CellComm.R
\name{interactionmatrix}
\alias{interactionmatrix}
\title{Compute the cell-cell interaction data frame.}
\usage{
interactionmatrix(result)
}
\arguments{
\item{result}{data frame; the interactions for each pair, plus the
pvalue column, as calculated by the computePvalue function.}
}
\value{
data frame;
}
\description{
Create a data frame with putative cell-cell interactions.
}
