% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{modelClustering}
\alias{modelClustering}
\title{Model-based clustering using the Mclust package.}
\usage{
modelClustering(object, assay.type = "RNA", num.pcs)
}
\arguments{
\item{object}{CellRouter object}

\item{assay.type}{character; the type of data to use.}

\item{num.pcs}{numeric; number of principal components that will define the
space used as input to perform model-based clustering.}
}
\description{
Perform model-based clustering and updates sampTab.
}
