% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper_functions.R
\name{multiplot}
\alias{multiplot}
\title{Combine multiple plots into a single one.}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot object; the plots.}

\item{plotlist}{list; the plots.}

\item{file}{character; filename where the plots will be saved.}

\item{cols}{numeric; number of columns in the generated figure.}

\item{layout}{matrix; A matrix specifying the layout. If present, 'cols' is
ignored.}
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot
objects). If the layout is something like matrix(c(1,2,3,3), nrow = 2,
byrow = TRUE), then plot 1 will go in the upper left, 2 will go in the upper
right, and 3 will go all the way across the bottom. Helper function of
plotClusterHeatmap and plotClusters.
}
