% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{nodeLabels}
\alias{nodeLabels}
\title{Assign labels for subpopulations.}
\usage{
nodeLabels(sampTab, column)
}
\arguments{
\item{sampTab}{data frame; metadata information.}

\item{column}{character; column corresponding to subpopulation.}
}
\description{
Assign labels for (for plotting). Found in graphClustering. Helper function
of findClusters.
}
