% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_method.R
\docType{methods}
\name{Normalize}
\alias{Normalize}
\alias{Normalize,CellRouter-method}
\title{Normalize the data.}
\usage{
Normalize(object, assay.type = "RNA")

\S4method{Normalize}{CellRouter}(object, assay.type = "RNA")
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}
}
\value{
CellRouter object with the ndata slot updated.
}
\description{
Normalize the data from CellRouter.
}
