% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTrajectories_method.R
\name{pathsinfo}
\alias{pathsinfo}
\title{Helper function of processTrajectories.}
\usage{
pathsinfo(
  object,
  assay.type = "RNA",
  num.cells,
  neighs,
  column.ann = "community",
  column.color = "colors"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{num.cells}{numeric; number of cells.}

\item{neighs}{numeric; neighs.}

\item{column.ann}{character; column corresponding to the population.}

\item{column.color}{character; column corresponding to the color.}
}
\value{
list; list with distr, path, pathInfo, pseudotime for each path, and
path_data.
}
\description{
Get information about the paths between populations.
}
