% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayenrichment.R
\docType{methods}
\name{pathwayenrichment}
\alias{pathwayenrichment}
\alias{pathwayenrichment,CellRouter-method}
\title{Pathway enrichment analysis}
\usage{
pathwayenrichment(
  object,
  names,
  cc = NULL,
  species = c("Hs", "Mm"),
  annotation,
  ids
)

\S4method{pathwayenrichment}{CellRouter}(
  object,
  names,
  cc = NULL,
  species = c("Hs", "Mm"),
  annotation,
  ids
)
}
\arguments{
\item{object}{CellRouter object.}

\item{names}{character vector; selected trajectories.}

\item{cc}{character vector; list of cell cycle genes or other gene lists to
be removed from the genes up or down-regulated in each trajectory.}

\item{species}{character; species: Hs for Homo Sapiens or Mm for Mus Musculus.}

\item{annotation}{character; organism-specific annotations: 'mouse' or
'org.Mm.eg.db', or 'human' or 'org.Hs.eg.db'.}

\item{ids}{table containing mappings between gene identifiers.}
}
\value{
CellRouter object with the pathwayenrichment slot updated.
}
\description{
Pathway enrichment analysis on genes up or down-regulated along trajectories.
}
