% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbranch_method.R
\docType{methods}
\name{plotbranch}
\alias{plotbranch}
\alias{plotbranch,CellRouter-method}
\title{Plot branch-specific kinetic patterns.}
\usage{
plotbranch(object, direction = c("up", "down"), p1, p2)

\S4method{plotbranch}{CellRouter}(object, direction = c("up", "down"), p1, p2)
}
\arguments{
\item{object}{CellRouter object.}

\item{direction}{character; plot genes up or down-regulated along
trajectories.}

\item{p1}{character; trajectory 1.}

\item{p2}{character; trajectory 2.}
}
\value{
list; ggplot2 graphs.
}
\description{
Plot z-scores across specific trajectories.
}
