% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterHeatmap_method.R
\docType{methods}
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\alias{plotClusterHeatmap,CellRouter-method}
\title{Plot cluster heatmap.}
\usage{
plotClusterHeatmap(object, show)

\S4method{plotClusterHeatmap}{CellRouter}(object, show)
}
\arguments{
\item{object}{CellRouter object.}

\item{show}{character vector; selected trajectories.}
}
\value{
list; ggplot2 plots.
}
\description{
Plot mean kinetic profile for each transcriptional cluster and save plots to
file.
}
