% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterProfiler_CellComm.R
\name{plotClusterProfiler}
\alias{plotClusterProfiler}
\title{Plot pathway enrichment analysis results.}
\usage{
plotClusterProfiler(
  table,
  num.pathways = 10,
  number.words = 3,
  order,
  flip = TRUE
)
}
\arguments{
\item{table}{compareClusterResult; the compareClusterResult of the
compareClusterResult class as obtained from the
clusterProfiler::compareCluster constructor.}

\item{num.pathways}{numeric; the maximum number of pathways considered in the
analysis.}

\item{number.words}{numeric; the minimum number of words for the cluster to
be considered.}

\item{order}{character vector; the ordered names of clusters.}

\item{flip}{boolean; apply coordinate flip (horizontal and vertical) to the
plot.}
}
\value{
ggplot2; plot.
}
\description{
Plot the cluster profiler and save to file.
}
