% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusters_method.R
\docType{methods}
\name{plotClusters}
\alias{plotClusters}
\alias{plotClusters,CellRouter-method}
\title{Plot kinetic trends for genes in transcriptional clusters.}
\usage{
plotClusters(object, p)

\S4method{plotClusters}{CellRouter}(object, p)
}
\arguments{
\item{object}{CellRouter object.}

\item{p}{character; selected trajectory.}
}
\value{
list; ggplot2 plots.
}
\description{
Plot the gene dynamics for each cluster in the selected trajectory and save
to file.
}
