% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDRExpression_method.R
\docType{methods}
\name{plotDRExpression}
\alias{plotDRExpression}
\alias{plotDRExpression,CellRouter-method}
\title{Show gene expression in the space of reduced dimensionality.}
\usage{
plotDRExpression(
  object,
  assay.type = "RNA",
  genelist,
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  threshold = 2,
  dims.use = c(1, 2),
  columns = 5,
  dotsize = 1,
  alpha = 0.5,
  title
)

\S4method{plotDRExpression}{CellRouter}(
  object,
  assay.type = "RNA",
  genelist,
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  threshold = 2,
  dims.use = c(1, 2),
  columns = 5,
  dotsize = 1,
  alpha = 0.5,
  title
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genelist}{character vector; genes to show.}

\item{reduction.type}{character; the dimension reduction space to be used:
pca, tsne, DC of diffusion components, umap, or custom.}

\item{threshold}{numeric; threshold to rescale gene expression.}

\item{dims.use}{numeric vector; dimensions to use.}

\item{columns}{numeric; number of columns in the output figure.}

\item{dotsize}{numeric; dot size.}

\item{alpha}{numeric; transparency (between 0 and 1).}

\item{title}{character; the title of the plot.}
}
\value{
ggplot2; plot.
}
\description{
Plot selected gene expression in reduced dimension space and save to file.
}
