% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichR.R
\name{plotEnrichR}
\alias{plotEnrichR}
\title{Plot Enrichment Result}
\usage{
plotEnrichR(
  enrichments,
  annotation = "Reactome_2016",
  num.pathways = 10,
  order
)
}
\arguments{
\item{enrichments}{list; enrichments.}

\item{annotation}{character; annotation of the enrichment.}

\item{num.pathways}{numeric; the number of pathways considered in the
enrichments.}

\item{order}{character vector; ordered populations.}
}
\value{
ggplot2 graph.
}
\description{
Create enrichment results plot.
}
