% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenesInClusters.R
\docType{methods}
\name{plotGenesInClusters}
\alias{plotGenesInClusters}
\alias{plotGenesInClusters,list-method}
\title{Plot top-ranked genes in transcriptional clusters.}
\usage{
plotGenesInClusters(assay.type = "RNA", rankedDynamics, traj.name, num.genes)

\S4method{plotGenesInClusters}{list}(assay.type = "RNA", rankedDynamics, traj.name, num.genes)
}
\arguments{
\item{assay.type}{character; the type of data to use.}

\item{rankedDynamics}{list; ranked dynamics as calculated by the
rankGenesTranscriptionalClusters function.}

\item{traj.name}{character; trajectory of interest.}

\item{num.genes}{numeric; number of genes to plot.}
}
\value{
ggplot2 graph.
}
\description{
Plot top-ranked genes calculated by the rankGenesTranscriptionalClusters
function and saves to file.
}
