% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{plotIntegration}
\alias{plotIntegration}
\title{Plot spatial clusters}
\usage{
plotIntegration(
  object,
  assay.type = "RNA",
  sample.name = "Sample1",
  cluster.column,
  colors.column,
  library.column,
  libraries
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.column}{character; the name of the column where the clustering
information is stored.}

\item{colors.column}{character; the name of the column where the color of each
cluster is stored.}

\item{library.column}{character; the name of the column where the library
information is stored.}

\item{libraries}{character; the selected library.}
}
\value{
ggplot2; the plots for all libraries.
}
\description{
Plot clusters in tissue image.
}
