% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteraction_CellComm.R
\name{plotInteractionMatrix2}
\alias{plotInteractionMatrix2}
\title{Plot inferred cell-cell interaction network.}
\usage{
plotInteractionMatrix2(graph)
}
\arguments{
\item{graph}{list; the graph and the pairs as generated by the cellnetwork3
function.}
}
\value{
list; the data frame which the rows are the cells of first type,
while the columns are the cells of the second type, and the plot. The values
of the data frame are the number of unique interactions between the pairs.
}
\description{
Plot the heatmap of the inferred cell-cell interaction network and return it.
The interaction data frame has cells of the first type as rows and cells of
the second type as columns, the values are the number of unique interactions
between the cell type pairs.
}
