% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteraction_CellComm.R
\name{plotInteractionNetwork}
\alias{plotInteractionNetwork}
\title{Plot graph of interactions.}
\usage{
plotInteractionNetwork(matrix, interactionList, interactionTable)
}
\arguments{
\item{matrix}{data frame; the interaction matrix as calculated by the
interactionmatrix function.}

\item{interactionList}{character vector; the list of types considered in the
interactions.}

\item{interactionTable}{matrix?; the interactions for each pair, plus the
pvalue column, as calculated bu the calculatePvalue function.}
}
\value{
list; the igraph interaction network and the plot.
}
\description{
Create a plot from the graph of interactions and save it to file.
}
