% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairDotplot_CellComm.R
\name{plotPairDotplot}
\alias{plotPairDotplot}
\title{Plot pair dotplot.}
\usage{
plotPairDotplot(tmp4, interactionList, num.pairs = 10)
}
\arguments{
\item{tmp4}{matrix?; the inferred cell-cell interactions and ligand-receptor
pairs.}

\item{interactionList}{character vector; the list of cell-cell interactions
to be plotted.}

\item{num.pairs}{numeric; the number of pairs that will be selected.}
}
\value{
list; the selected interactions matrix and the plot.
}
\description{
Plot the mean expression dotplot for the specified ligand-receptor pairs.
}
