% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPathHeatmap_method.R
\docType{methods}
\name{plotPathHeatmap}
\alias{plotPathHeatmap}
\alias{plotPathHeatmap,CellRouter-method}
\title{Plot path heatmap.}
\usage{
plotPathHeatmap(
  object,
  assay.type = "RNA",
  paths,
  genelist,
  cluster.column = "population",
  color.column = "colors",
  threshold = 2
)

\S4method{plotPathHeatmap}{CellRouter}(
  object,
  assay.type = "RNA",
  paths,
  genelist,
  cluster.column = "population",
  color.column = "colors",
  threshold = 2
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{paths}{character vector; selected trajectories.}

\item{genelist}{character vector; genes to show.}

\item{cluster.column}{character; the name of the column where the clustering
information is stored.}

\item{color.column}{character; the name of the column where the color
information is stored.}

\item{threshold}{numeric; threshold to rescale gene expression.}
}
\value{
list; ggplot2 plots.
}
\description{
Plot heatmap for specified genes in the trajectories and save plots to
directory.
}
