% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportion.R
\name{plotProportion2}
\alias{plotProportion2}
\title{Proportion plot.}
\usage{
plotProportion2(
  object,
  assay.type = "RNA",
  condition,
  population,
  color.column,
  order = NULL
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{condition}{character; column in the metadata table specifying an
annotation, such as sorted populations.}

\item{population}{character; column in the metadata table specifying another
annotation.}

\item{color.column}{character; corresponding color column for the annotation.}

\item{order}{character vector; ordered populations. The default is ascending
order.}
}
\value{
list; the data frame of cells (condition) and their classifications
(population), and the plot.
}
\description{
Plot a graph of the cells in two populations.
}
