% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDimension.R
\name{plotReducedDimension}
\alias{plotReducedDimension}
\title{Plot reduced dimension space.}
\usage{
plotReducedDimension(
  object,
  assay.type = "RNA",
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  dims.use = c(1, 2),
  annotation,
  annotation.color,
  showlabels,
  dotsize = 1,
  labelsize = 3,
  convex = FALSE,
  percentage = 0.8,
  alpha = 0.1
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{reduction.type}{character; the dimension reduction space to be used:
pca, tsne, DC of diffusion components, umap, or custom.}

\item{dims.use}{numeric vector; dimensions to use.}

\item{annotation}{character; column in the metadata table to annotate the figure.}

\item{annotation.color}{character; corresponding color column for the annotation.}

\item{showlabels}{logical; show labels in the plot.}

\item{dotsize}{numeric; dot size.}

\item{labelsize}{numeric; label size.}

\item{convex}{boolean; whether to connect the dots.}

\item{percentage}{numeric; proportion of data in the panel (between 0 and 1).}

\item{alpha}{numeric; transparency (between 0 and 1).}
}
\value{
ggplot2; plot.
}
\description{
Plot reduced dimension space using pca, tsne, diffusion components, UMAP or
custom.
}
