% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegulonDynamics.R
\docType{methods}
\name{plotRegulonDynamics}
\alias{plotRegulonDynamics}
\alias{plotRegulonDynamics,CellRouter-method}
\title{Plot regulon dynamics.}
\usage{
plotRegulonDynamics(object, p, regulator, scores, cluster = TRUE)

\S4method{plotRegulonDynamics}{CellRouter}(object, p, regulator, scores, cluster = TRUE)
}
\arguments{
\item{object}{CellRouter object.}

\item{p}{character; selected trajectory.}

\item{regulator}{character; the selected regulator.}

\item{scores}{list; scores of transcriptional regulators.}

\item{cluster}{boolean; whether to cluster kinetic patterns.}
}
\value{
list; ggplot2 plots.
}
\description{
Plot derivative and kinetic patterns of predicted regulators of cell-fate
transitions and save to file.
}
