% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignatureScore.R
\docType{methods}
\name{plotSignatureScore}
\alias{plotSignatureScore}
\alias{plotSignatureScore,CellRouter-method}
\title{Show gene expression in the space of reduced dimensionality.}
\usage{
plotSignatureScore(
  object,
  assay.type = "RNA",
  genelist,
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  threshold = 2,
  dims.use = c(1, 2),
  columns = 5,
  dotsize = 1,
  alpha = 0.5
)

\S4method{plotSignatureScore}{CellRouter}(
  object,
  assay.type = "RNA",
  genelist,
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  threshold = 2,
  dims.use = c(1, 2),
  columns = 5,
  dotsize = 1,
  alpha = 0.5
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genelist}{character vector; genes to show.}

\item{reduction.type}{character; the dimension reduction space to be used:
pca, tsne, DC of diffusion components, umap, or custom.}

\item{threshold}{numeric; threshold to rescale gene expression.}

\item{dims.use}{numeric vector; dimensions to use.}

\item{columns}{numeric; number of columns in the output figure.}

\item{dotsize}{numeric; dot size.}

\item{alpha}{numeric; transparency (between 0 and 1).}
}
\value{
ggplot2; plot.
}
\description{
Show gene expression in the selected space of reduced dimensionality and save
to file.
}
