% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignatureScoreDynamics.R
\docType{methods}
\name{plotSignatureScoreDynamics}
\alias{plotSignatureScoreDynamics}
\alias{plotSignatureScoreDynamics,CellRouter-method}
\title{Plot signature scores along a trajectory.}
\usage{
plotSignatureScoreDynamics(
  object,
  assay.type = "RNA",
  trajectories,
  geneList,
  rescale,
  order = 1
)

\S4method{plotSignatureScoreDynamics}{CellRouter}(
  object,
  assay.type = "RNA",
  trajectories,
  geneList,
  rescale,
  order = 1
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{trajectories}{character vector; selected trajectories.}

\item{geneList}{character vector; genes to show.}

\item{rescale}{boolean; whether to rescale curves simultaneously or
individually.}

\item{order}{numeric;     integer giving the order of the neighborhood.}
}
\value{
ggplot2 graph.
}
\description{
Create trajectory signature scores plot.
}
