% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSignaturesHeatmap.R
\name{plotSignaturesHeatmap}
\alias{plotSignaturesHeatmap}
\title{Plot heatmap with gene signatures.}
\usage{
plotSignaturesHeatmap(
  object,
  assay.type = "RNA",
  markers,
  column.ann,
  column.color,
  num.cells = NULL,
  threshold = 2,
  genes.show = NULL,
  low = "purple",
  intermediate = "black",
  high = "yellow",
  order = NULL,
  fontsize = 5
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{markers}{tibble; genes preferentially expressed in each column.ann, as
obtained in the findSignatures function. For example, in clusters or sorted
populations.}

\item{column.ann}{character; column in the metadata table used to annotate
the kNN graph. For example, clusters, sorted cell populations.}

\item{column.color}{character; column corresponding to the colors.}

\item{num.cells}{numeric; number of cells to show in the heatmap.}

\item{threshold}{numeric; threshold used to center the data.}

\item{genes.show}{character vector;  gene names to show in the heatmap. The
default is the top 5 genes according to fold change for each population.}

\item{low}{character; color for low expression.}

\item{intermediate}{character; color for intermediate expression.}

\item{high}{character; color for high expression.}

\item{order}{character vector; order of population names, the default is
alphabetical order.}

\item{fontsize}{numeric; font size.}
}
\value{
pheatmap; plot.
}
\description{
Plot heatmap of gene expression for the specified genes in each population.
The plot is not shown in the RStudio Plots display, but it is saved in the
specified directory.
}
