% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\name{plotSpatialClusters}
\alias{plotSpatialClusters}
\title{Plot spatial clusters}
\usage{
plotSpatialClusters(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column = "population",
  colors.column = "colors",
  annotate_centroids = FALSE,
  point.size = 1.75,
  title = NULL
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.column}{character; the name of the column where the clustering
information is stored.}

\item{colors.column}{character; the name of the column where the color of each
cluster is stored.}

\item{annotate_centroids}{boolean; annotate the cluster centroids.}

\item{point.size}{numeric; the size of the spot in the image.}

\item{title}{character; figure title.}
}
\value{
ggplot2; plot.
}
\description{
Plot clusters in tissue image.
}
