% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\name{plotSpatialExpression}
\alias{plotSpatialExpression}
\title{Plot spatial expression}
\usage{
plotSpatialExpression(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column = "population",
  selected_clusters = NULL,
  genes,
  point.size = 1.75
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.column}{character; the name of the column where the
clustering information is stored.}

\item{selected_clusters}{list; a vector containing the name of the clusters
that will be plotted for each tissue sample.}

\item{genes}{character; a vector containing the name of the genes.
that will be plotted for each tissue sample.}

\item{point.size}{numeric; the size of the spot in the image.}
}
\value{
list; ggplot2 plots.
}
\description{
Plot spatial expression for selected genes.
}
