% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\name{plotSpatialSubclusters}
\alias{plotSpatialSubclusters}
\title{Plot subclusters}
\usage{
plotSpatialSubclusters(
  object,
  assay.type = "ST",
  sample.name = "Sample1",
  cluster.column = "population",
  subcluster.column = "Subpopulation",
  selected_clusters,
  annotate_centroids = FALSE,
  point.size = 1.75,
  title = NULL
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.column}{character; the name of the column where the
clustering information is stored.}

\item{subcluster.column}{character; the name of the column where the
subclustering information is stored.}

\item{selected_clusters}{list; a vector containing the name of the clusters
that will be plotted for each tissue sample.}

\item{annotate_centroids}{boolean; annotate the subcluster centroids.}

\item{point.size}{numeric; the size of the spot in the image.}

\item{title}{character; figure title.}
}
\value{
ggplot2; plot.
}
\description{
Plot subclusters from selected clusters.
}
