% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottr_method.R
\docType{methods}
\name{plottr}
\alias{plottr}
\alias{plottr,CellRouter-method}
\title{Plot derivative and kinetic patterns of predicted regulators.}
\usage{
plottr(object, p, scores, cluster = TRUE)

\S4method{plottr}{CellRouter}(object, p, scores, cluster = TRUE)
}
\arguments{
\item{object}{CellRouter object.}

\item{p}{character; trajectory.}

\item{scores}{numeric vector; scores of transcriptional regulators.}

\item{cluster}{boolean; whether to cluster kinetic patterns.}
}
\value{
list; ggplot2 plots.
}
\description{
Plot dynamics and derivative plots of transcriptional regulators of cell-fate
transitions based on the GRN score.
}
