% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottrajectories_method.R
\docType{methods}
\name{plottrajectories}
\alias{plottrajectories}
\alias{plottrajectories,CellRouter-method}
\title{Plot genes along specified trajectories.}
\usage{
plottrajectories(object, trajectories, geneList, rescale, columns = 5)

\S4method{plottrajectories}{CellRouter}(object, trajectories, geneList, rescale, columns = 5)
}
\arguments{
\item{object}{CellRouter object.}

\item{trajectories}{character vector; selected trajectories.}

\item{geneList}{character vector; genes to show.}

\item{rescale}{boolean; whether to rescale curves simultaneously or
individually.}

\item{columns}{numeric; the number of columns in the output figure.}
}
\value{
ggplot2; plot.
}
\description{
Plot gene expressions along different trajectories and save to file.
}
