% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottrajectories_method.R
\docType{methods}
\name{plottrajectory}
\alias{plottrajectory}
\alias{plottrajectory,CellRouter-method}
\title{Plot genes along a trajectory.}
\usage{
plottrajectory(object, assay.type = "RNA", trajectory, geneList)

\S4method{plottrajectory}{CellRouter}(object, assay.type = "RNA", trajectory, geneList)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{trajectory}{character; selected trajectory.}

\item{geneList}{character vector; gene list.}
}
\value{
ggplot; the graph.
}
\description{
Plot gene expressions along a trajectory, return ggplot and save to file.
}
