% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotValue.R
\docType{methods}
\name{plotValue}
\alias{plotValue}
\alias{plotValue,CellRouter-method}
\title{Show any numeric value in the CellRouter metadata table in the space of reduced dimensionality.}
\usage{
plotValue(
  object,
  assay.type = "RNA",
  variable.show,
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  threshold = 2,
  dims.use = c(1, 2),
  columns = 5,
  dotsize = 1,
  alpha = 0.5
)

\S4method{plotValue}{CellRouter}(
  object,
  assay.type = "RNA",
  variable.show,
  reduction.type = c("tsne", "pca", "DC", "umap", "custom"),
  threshold = 2,
  dims.use = c(1, 2),
  columns = 5,
  dotsize = 1,
  alpha = 0.5
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{variable.show}{character vector; values to show.}

\item{reduction.type}{character; the dimension reduction space to be used:
pca, tsne, DC of diffusion components, umap, or custom.}

\item{threshold}{numeric; threshold to rescale gene expression.}

\item{dims.use}{numeric vector; the number of dimensions to use.}

\item{columns}{numeric; number of columns in the output figure.}

\item{dotsize}{numeric; dot size.}

\item{alpha}{numeric; transparency (between 0 and 1).}
}
\value{
ggplot2; plot.
}
\description{
Show any numeric value in the CellRouter metadata table (number of detected
genes, etc.) in the selected space of reduced dimensionality and save
to file.
}
