% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Plot violin plot.}
\usage{
plotViolin(
  object,
  assay.type = "RNA",
  geneList,
  column,
  column.color,
  cols,
  order = NULL,
  log = TRUE
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{geneList}{character vector; genes to show.}

\item{column}{character; column in the metadata table to annotate the figure.}

\item{column.color}{character; corresponding color column for the annotation.}

\item{cols}{numeric; the number of columns in the output figure.}

\item{order}{character vector; ordered populations. The default is ascending
order.}

\item{log}{boolean; whether to use data in logarithmic scale.}
}
\value{
ggplot2; plot.
}
\description{
Plot violin plot of gene expression in the populations and save to file.
}
