% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.pairing_CellComm.R
\name{population.pairing}
\alias{population.pairing}
\title{Compute cell-cell interaction network.}
\usage{
population.pairing(mean.expr, ligands, receptors, threshold, pairs)
}
\arguments{
\item{mean.expr}{list; the statistics of the expressed genes as calculated by
the computeValue method of CellRouter.}

\item{ligands}{vector; list of genes annotated as ligands.}

\item{receptors}{vector; list of genes annotated as receptors.}

\item{threshold}{numeric; threshold to select the genes in the population
according to their mean expression.}

\item{pairs}{the pairs;}
}
\value{
data frame; interactions with celltype1, celltype2, pair, ligand,
and receptor.
}
\description{
Analyze the mean expression data generated by the computeValue method of
CellRouter to determine co-expression patterns of ligands and receptors.
}
