% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCell.R
\name{predictCellCycle}
\alias{predictCellCycle}
\title{Predict cell cycle phase.}
\usage{
predictCellCycle(object, assay.type = "RNA", columns)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{columns}{character vector; columns to be selected from the metadata
table.}
}
\value{
CellRouter object with the sampTab slot updated.
}
\description{
Predict cell cycle phase based on scores from the metadata table and add
phase column with the results to it. Call the addInfo function to add the
information to assay@sampTab.
}
