% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellComm_functions.R
\name{predictCellInteractions}
\alias{predictCellInteractions}
\title{Predict Cell Interactions}
\usage{
predictCellInteractions(
  object,
  assay.type = "RNA",
  sample.name = "Sample1",
  cluster.type = "Cluster",
  graph,
  distance.threshold = 0.75
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{cluster.type}{character; the column where the clusters are
indicated. It could be either 'Cluster' or 'Subcluster'. Used to locate the
distance matrix in the CellRouter object.}

\item{graph}{igraph; the cell graph.}

\item{distance.threshold}{numeric; the threshold for the distance.}
}
\value{
list of ggplot2 graphs.
}
\description{
Plot graphs of the ligands and receptors, their distances and scores.
}
