% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCell.R
\name{predictState}
\alias{predictState}
\title{Predict state based on scored gene lists.}
\usage{
predictState(object, assay.type = "RNA", columns, col.name)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{columns}{character vector; columns to be selected from the metadata
table.}

\item{col.name}{character; column name to be added to the metadata table
after state prediction.}
}
\value{
CellRouter object with the sampTab slot updated.
}
\description{
Predict cell cycle phase based on scores from the metadata table and add
column with the results to it. Call the addInfo function to add the
information to assay@sampTab.
}
