% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCurves.R
\docType{methods}
\name{processCurves}
\alias{processCurves}
\alias{processCurves,CellRouter-method}
\title{Process curves}
\usage{
processCurves(
  object,
  assay.type = "RNA",
  genes,
  num.cells,
  neighs,
  column.ann = "population",
  column.color = "colors"
)

\S4method{processCurves}{CellRouter}(
  object,
  assay.type = "RNA",
  genes,
  num.cells,
  neighs,
  column.ann = "population",
  column.color = "colors"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes}{character vector; gene names that are used for trajectory
analysis.}

\item{num.cells}{numeric; minimum of cells contained in the trajectories.}

\item{neighs}{numeric; the size of the neighborhood in kNN graph used to
smoothen kinetic profiles.}

\item{column.ann}{character; transitions between the cell states provided
here are identified, such as clusters or sorted populations.}

\item{column.color}{character; the colors corresponding to the annotation
in column.ann.}
}
\value{
CellRouter object with the curves and the pathsinfo slots updated.
}
\description{
Process curves found in findCurves.
}
