% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTrajectories_method.R
\docType{methods}
\name{processTrajectories}
\alias{processTrajectories}
\alias{processTrajectories,CellRouter-method}
\title{Process trajectories.}
\usage{
processTrajectories(
  object,
  assay.type = "RNA",
  genes,
  path.rank,
  num.cells,
  neighs,
  column.ann = "population",
  column.color = "colors"
)

\S4method{processTrajectories}{CellRouter}(
  object,
  assay.type = "RNA",
  genes,
  path.rank,
  num.cells,
  neighs,
  column.ann = "population",
  column.color = "colors"
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes}{character vector; gene names that are used for trajectory
analysis.}

\item{path.rank}{character; method to rank paths: path_cost, path_flow, rank,
length. Available ranks are 'path_cost', 'path_flow', 'rank', 'length'.}

\item{num.cells}{numeric; minimum of cells contained in the trajectories.}

\item{neighs}{numeric; the size of the neighborhood in kNN graph used to
smoothen kinetic profiles.}

\item{column.ann}{character; transitions between the cell states provided
here are identified, such as clusters or sorted populations.}

\item{column.color}{character; the colors corresponding to the annotation
in column.ann.}
}
\value{
CellRouter object with the paths, the networks and the pathsinfo
slots updated.
}
\description{
Process trajectories found in findPaths.
}
