% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cRamp_helper.R
\name{range01}
\alias{range01}
\title{Rescale x from 0 to 1.}
\usage{
range01(x)
}
\arguments{
\item{x}{values to be converted to colors.}
}
\description{
Helper function. Rescale x from 0 to 1 by subtracting the minimum value and
dividing by the range.
}
