% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSignatures.R
\name{ranked_findSpecGenes}
\alias{ranked_findSpecGenes}
\title{Helper function of findSignatures.}
\usage{
ranked_findSpecGenes(
  expDat,
  sampTab,
  qtile = 0.95,
  remove = FALSE,
  dLevel = "population_name"
)
}
\arguments{
\item{expDat}{Expression matrix}

\item{sampTab}{Sample annotation table}

\item{qtile}{numeric; quantile (between 0 and 1) to select top genes
correlated with the idealized expression pattern.}

\item{remove}{logical; remove overlaping genes.}

\item{dLevel}{character; annotation level to group on, groups to compare.}
}
\value{
; specificSets.
}
\description{
Find gene signatures based on a template-matching approach.
Find genes that are preferentially expressed in specified samples. Find it by
adjusting a line to the gene expression for each cell population, the getting
the most representative genes, the ones with better fit.
}
