% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankGenesTranscriptionalClusters.R
\docType{methods}
\name{rankGenesTranscriptionalClusters}
\alias{rankGenesTranscriptionalClusters}
\alias{rankGenesTranscriptionalClusters,CellRouter-method}
\title{Rank genes in transcriptional clusters.}
\usage{
rankGenesTranscriptionalClusters(object, num.genes)

\S4method{rankGenesTranscriptionalClusters}{CellRouter}(object, num.genes)
}
\arguments{
\item{object}{CellRouter object.}

\item{num.genes}{numeric; top num.genes in each transcriptional cluster.}
}
\value{
list; cluster trends for each trajectory.
}
\description{
Rank genes in each transcriptional cluster based on their correlation with
the mean kinetic profile.
}
