% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankpaths_CellComm.R
\name{rankpaths}
\alias{rankpaths}
\title{Rank signaling pathways.}
\usage{
rankpaths(summary, apathways)
}
\arguments{
\item{summary}{data frame; the summary of the files obtained in the
findpaths.simple function.}

\item{apathways}{list; the pathways as calculated by the activeSignaling
function.}
}
\value{
list; the list of ranked paths for each interaction (npath) and the
list with the number of targets in it (n.targets).
}
\description{
Rank signaling pathways based on the number of targets, the rank, and
the mean flow.
}
