% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\name{read10X}
\alias{read10X}
\title{Read 10X expression data.}
\usage{
read10X(matrix_dir)
}
\arguments{
\item{matrix_dir}{character; the path to the directory with the barcodes,
features, and matrix files. The string should NOT end with your system's path
separator.}
}
\description{
Read the spatial transcriptomics expression matrix given a directory with
the barcodes (barcodes.tsv.gz), features (features.tsv.gz), and matrix
(matrix.mtx.gz) files.
}
