% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_transcriptomics.R
\docType{methods}
\name{read10XImage}
\alias{read10XImage}
\alias{read10XImage,CellRouter-method}
\title{Read ST metadata.}
\usage{
read10XImage(
  object,
  assay.type = "ST",
  sample_names,
  image_paths,
  scalefactor_paths,
  tissue_paths
)

\S4method{read10XImage}{CellRouter}(
  object,
  assay.type = "ST",
  sample_names,
  image_paths,
  scalefactor_paths,
  tissue_paths
)
}
\arguments{
\item{object}{the CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample_names}{character; names of the tissue samples.}

\item{image_paths}{character; file path of the tissue image (png format).}

\item{scalefactor_paths}{character; file path of the scale factor (json
format).}

\item{tissue_paths}{character; file path of the positions of tissue barcodes
(csv format).}
}
\value{
the CellRouter object.
}
\description{
Read the tissue image, the scale factor and the barcodes of the spatial
transcriptomics metadata and add them to the respective assay metadata.
}
