% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatornetwork.R
\name{regulatornetwork}
\alias{regulatornetwork}
\title{Plot subnetworks for selected transcriptional regulators in the trajectories.}
\usage{
regulatornetwork(object, x, ggrn, genelist)
}
\arguments{
\item{object}{CellRouter object.}

\item{x}{list; GRN scores as calculated by the grnscores function.}

\item{ggrn}{igraph; the gene regulatory network.}

\item{genelist}{character vector; gene names containing the transcriptional
regulators from which the network modules will be identified.}
}
\value{
list; rgrn graphs for all names in x, and plot.
}
\description{
In development, currently not working.
Plot a network module enriched around transcriptional regulators of
cell-fate transitions.
}
