% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparse_matrix.R
\name{scale.sparse}
\alias{scale.sparse}
\title{Scale sparse matrix dividing it in blocks.}
\usage{
\method{scale}{sparse}(matrix, blocksize = 500)
}
\arguments{
\item{matrix}{any matix; the data tha will be scaled.}

\item{blocksize}{numeric; the size of the blocks used in each scaling.}
}
\value{
any matrix; the scaled matrix.
}
\description{
An alternative version of scale for sparse matrix. Avoid errors for large
matrixes.
}
