% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\docType{methods}
\name{scaleData}
\alias{scaleData}
\alias{scaleData,CellRouter-method}
\title{Scale data}
\usage{
scaleData(
  object,
  assay.type = "RNA",
  genes.use = NULL,
  scale.max = 10,
  vars.regress = NULL,
  blocksize = 500
)

\S4method{scaleData}{CellRouter}(
  object,
  assay.type = "RNA",
  genes.use = NULL,
  scale.max = 10,
  vars.regress = NULL,
  blocksize = 500
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{genes.use}{vector; genes to scale/center. Default is all genes in
normalized data.}

\item{scale.max}{numeric; max value in scaled data.}

\item{vars.regress}{character vector; variables to regress out.}

\item{blocksize}{numeric; size of the blocks in which genes will be scaled.}
}
\value{
CellRouter object with the scale.data slot updated.
}
\description{
Scale and center the data. Individually regress variables provided in
vars.regress using using a linear model. Other models are under development.
}
