% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreGeneSets.R
\name{scoreGeneSets}
\alias{scoreGeneSets}
\title{Score gene sets.}
\usage{
scoreGeneSets(
  object,
  assay.type = "RNA",
  sample.name = "Sample1",
  genes.list,
  bins = 25,
  genes.combine = NULL
)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{sample.name}{character; the name of the tissue sample.}

\item{genes.list}{list; gene lists for which the scores will be calculated.}

\item{bins}{numeric; number of bins to split expression data.}

\item{genes.combine}{character; genes to combine. The default is all genes in
normalized data.}
}
\value{
CellRouter object with the sampTab slot updated.
}
\description{
Description.
}
