% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaling2TF_CellComm.R
\name{signaling2TF}
\alias{signaling2TF}
\title{Signaling to transcription factors.}
\usage{
signaling2TF(npaths, interaction, num.pathways = 10)
}
\arguments{
\item{npaths}{list; the ranked pathways and the number of number of targets
in each interaction as calculated by the rankpaths function.}

\item{interaction}{character; the name of the selected interaction.}

\item{num.pathways}{numeric; the number of pathways considered in the
signaling.}
}
\value{
list; a path and interaction data frame, and the plot.
}
\description{
Select the paths in the specified interaction and plot the activity score.
Save the plot to file and return the data frame used to generate it.
}
