% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothDynamics_method.R
\docType{methods}
\name{smoothDynamics}
\alias{smoothDynamics}
\alias{smoothDynamics,CellRouter-method}
\title{Smooth dynamics along trajectories.}
\usage{
smoothDynamics(object, names)

\S4method{smoothDynamics}{CellRouter}(object, names)
}
\arguments{
\item{object}{CellRouter object.}

\item{names}{character vector; selected trajectories. Default consists of
all trajectories.}
}
\value{
CellRouter object with the dynamics slot updated.
}
\description{
Smooth the dynamics of gene expression by the path pseudotime using
polynomial regression. Helpful to cluster complex transcriptional patterns.
}
