% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcestargets_CellComm.R
\name{sourcestargets}
\alias{sourcestargets}
\title{Create source target list for each interaction.}
\usage{
sourcestargets(matrix, threshold, tfs)
}
\arguments{
\item{matrix}{data frame; the interaction matrix as calculated by the
interactionmatrix function.}

\item{threshold}{numeric; select the interactions which values are above the
threshold.}

\item{tfs}{character vector; the transcription factors.}
}
\value{
list; the receptor, the ligand, and interaction data frame for each
interaction.
}
\description{
For each interaction above the threshold, create a list of receptors,
containing cell surface receptor as sources and the transcriptional
regulators as targets; a list of ligands, sources as ligands and targets as
transcriptional regulators; and a data frame of ligands, receptors, and
interaction value.
}
