% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparse_matrix.R
\name{sparse.cor}
\alias{sparse.cor}
\title{Correlation function for sparse matrix.}
\usage{
sparse.cor(x)
}
\arguments{
\item{x}{matrix; data.}
}
\description{
Code found in \href{https://stackoverflow.com/questions/5888287/running-cor-or-any-variant-over-a-sparse-matrix-in-r}{Running cor() (or any variant) over a sparse matrix in R}.
}
