% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_sparse_matrix.R
\name{sparse.cor2}
\alias{sparse.cor2}
\title{Correlation function for sparse matrix.}
\usage{
sparse.cor2(x)
}
\arguments{
\item{x}{matrix; data.}
}
\description{
Alternative version of sparse.cor1.
}
