% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDimension.R
\docType{data}
\name{StatBag}
\alias{StatBag}
\title{Helper function of stat_bag.}
\format{
An object of class \code{Statbag} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatBag
}
\value{
ggproto object.
}
\description{
Creates ggproto object. Originally from aplpack package, plotting functions
removed.
}
\keyword{datasets}
