% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.flow_CellComm.R
\name{summarize.flow}
\alias{summarize.flow}
\title{Summarize flow network.}
\usage{
summarize.flow(files, prefix)
}
\arguments{
\item{files}{list; the filenames passed to the findpaths.simple function.}

\item{prefix}{character; the path to the files.}
}
\value{
list; the nodes, edges, paths and graphs (gml) for each celltype.
}
\description{
Resume the results of the findpaths.simple function.
}
