% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGenes_method.R
\docType{methods}
\name{topGenes}
\alias{topGenes}
\alias{topGenes,CellRouter-method}
\title{Find genes highly correlated with pseudotime.}
\usage{
topGenes(object, max.quantile, min.quantile)

\S4method{topGenes}{CellRouter}(object, max.quantile, min.quantile)
}
\arguments{
\item{object}{CellRouter object}

\item{max.quantile}{numeric; quantile to select positively correlated genes.}

\item{min.quantile}{numeric; quantile to select negatively correlated genes.}
}
\value{
CellRouter object with the top.correlations slot updated.
}
\description{
Find genes highly up-correlated or down-correlated genes in each path
according to the specified quantiles.
}
