% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCurves_method.R
\name{verify_gene_exp}
\alias{verify_gene_exp}
\title{Verify gene expression.}
\usage{
verify_gene_exp(object, assay.type = "RNA", curve_data, gene)
}
\arguments{
\item{object}{CellRouter object.}

\item{assay.type}{character; the type of data to use.}

\item{curve_data}{data frame; data from the curve including cell, dist and
clust columns.}

\item{gene}{string; gene name.}
}
\description{
Plot a gene expression across a curve.
}
